requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
streamlit==1.29.0
Pillow==10.1.0
scipy==1.11.4
